topology/filtration.h : 1. Replaced unsigned int -> unsigned long

+++ src/topology/filtration.h
@@ -92,7 +92,7 @@ class Filtration
         // Serialization
         friend class                            boost::serialization::access;
         template<class Archive>
-        void                                    serialize(Archive& ar, const unsigned int)
+        void                                    serialize(Archive& ar, const unsigned long)
         { ar & boost::serialization::make_nvp("order", container_); }
 };
